requirements.txt
asgiref==3.5.2
autopep8==1.6.0
boto3==1.24.22
botocore==1.27.22
Django==4.0.5
django-crispy-forms==1.14.0
django-environ==0.9.0
django-filter==22.1
django-localflavor==3.1
django-storages==1.12.3
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.1.1
psycopg2==2.9.3
pycodestyle==2.8.0
python-dateutil==2.8.2
python-stdnum==1.17
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
whitenoise==6.2.0